Severe bug fix.
As unimportant and innocent a bug as it looks,
it can cause massive destructions during automatic
installation of the profile (drush site-install btranslator ...)
And if you run the installation script as root (which I was
careless enough to do), it will erase the whole system.
And also it was hard to debug what was the problem.

I was burned by it, I hope that nobody else was damaged.
If yes, please accept my appologies.

The lesson learned is: always use chroot for new or
untested/unstable things
(http://joshrendek.com/2013/01/securing-ubuntu/#chroot)